fix(directus-next): no same-path page settings mismatches
