+++ libs/directus/directus-next/src/index.ts
@@ -1,7 +1,7 @@
 export * from './draft'
 export * from './redirect'
 export { logger as DirectusNextLogger } from './logger'
-export { directusRouteRouter } from './lib/directusRouteRouter'
+export { directusRouteRouter } from './router/router'
 export { getJsonErrorResponse } from './response'
 export * from './types'
 export type {
