+++ libs/directus/directus-next/src/server.ts
@@ -1,3 +1,3 @@
-export { directusRouteRouter } from './lib/directusRouteRouter'
+export { directusRouteRouter } from './router/router'
 export type { DirectusRouteConfig } from './types/directusRouteConfig'
 export * from './pageSettings'
